Add files via upload
This is the model trained on FER-2013 and muxspace's facial_expressions. Further research will address the following: anger(5,166), sad(6,278), surprised(4,264), happy(17,676), disgusted(649), neutral( 17,243), fearful(5,128), so the ratio of the number of images in the smallest set to the number of images in the largest set is 1:27. The screenshot added is the outcomes of training.